revert restructuring the components
