+++ src/components/main-page/Hero/Hero.module.scss
