+++ src/components/ui/Section/Section.module.scss
@@ -19,7 +19,7 @@
       top: 0;
       width: 100%;
       height: 100%;
-      background: url("../../assets/circle-pattern.svg");
+      background: url("../../../assets/circle-pattern.svg");
       background-repeat: repeat-y;
       background-position: 50% 0;
       opacity: 0.1;
